MEGA links reactions, first approach
- Added FileLinkViewController, now you can see a file public node. For
the moment you can download files to the App Sandbox.

- AppDelegate:
“application: openURL: sourceApplication: annotation:” MEGA links
reactions.
“applicationWillTerminate:” The unfinished downloads are deleted when
the application is killed.

- Helper:
Added pathForOffline and fileTypeIconExtension methods.

- OfflineTableViewController:
Now it's an Offline section.

- LoginViewController:
When a user is not logged, and clicks on a MEGA link, the app opens and
remember his choice (Import or Download) asks for login and the action
previously selected gets triggered.

- Main.storyboard, DetailsNodeInfoViewController,
CreateAccountViewController, ConfirmAccountViewController:
UI adjustments for iPhone 4S. Minor changes; some related to the
Offline section.

- SettingsViewController, TourViewController and main:
Minor changes.