Downloading and speed details. Cancel transfers
-AppDelegate:
Added a register of downloaded nodes on NSUserDefaults.

-Cloud.storyboard and CloudDriveTableViewController
Added variation of the current cell to show the speed and the
percentage completed of the node that is being downloaded.

-DetailsNodeInfoViewController
Added label with the average speed of the node download.

-NodeTableViewCell
Added function to cancel a download transfer.

-Added cancel icon, downloaded and downloading arrow images.

-FileLinkViewController, FolderLinkViewController.:
Deleted some unnecessary code.

-OfflineTableViewController:
Delete node from downloaded register when is deleted on Offline.

-Helper:
Workarounds for downloading and downloaded arrow images. Dictionaries
to have a temporary register when the app is running.