Fixes for login, fetch nodes process and Camera Uploads pop up
- AppDelegate:
Everything related to the login and fetch nodes process is now
centralized on the AppDelegate. The Camera Uploads pop up is managed
here too. The same for the process of file and folder links.

- CloudDriveTVC, ConfirmAccountVC and LoginVC:
Deleted everything related to the login and fetch nodes process, Camera
Uploads pop up and the process of file and folder links.

- FileLinkVC, FolderLinkVC and Helper:
Added a new way of managing the selected option (Import or download) on
file and folder links when you are not logged.

Also, added fix to avoid "Unbalanced calls to begin/end appearance
transitions for...".