Changed buttons and added some improvements on node details
Changed buttons, are now included on a table view with labels.
Improved “Save for Offline” functionality. Now a percentage and speed
label is shown when the transfer is active, like the Cloud Drive. Also,
you can cancel the transfer.
Now, the labels with long names are truncated on the middle not at the
end.