Bugs fixes
- AppDelegate:
Delete link after it has been opened with “Open in” functionality, to
avoid reprocessing.
Fixed bug on iOS 7 when you “Open in” and login (#817).
Improved dismissal of presented views when you open a link.

- OpenInActivity + DetailsNodeInfoViewController:
Finish activity and present on the completion handler the document
interaction controller to manage everything correctly on iOS 7.

- BrowserViewController:
Avoid another crash on iOS 7 presenting action sheet as needed (#839).