Share action instead of Get link on CloudDriveTableViewController, PhotosViewController and ShareItemsViewController
- Refactored all the code related with getting links or removing them,
request management to the AppDelegate, and UIActivityViewController to
the Helper for reusing it in the entire project.
- DetailsNodeInfoViewController, changed UIActivityViewController code
for the refactored one.
- Now GetLinkActivity, RemoveLinkActivity and ShareFolderActivity are
compatible with operations over several nodes. Some refactoring and
cleanup applied there.
- Make the get/remove link actions blocking while are being carried
out. Sharing actions is limited to 100 nodes tops. Avoid that the links
get copied on the pasteboard when you use MEGAActivityItemProvider.
- Fixed issues with iOS 7.
- Added new strings related to operations over several nodes.
- Removed unnecessary #import and reordered on some classes.