Save photos and videos directly in the gallery (#4005)
Downloads by defaults are saved in the Offline section.
Add two options in the advanced settings: "Save images to gallery" and "Save videos to gallery".
If those options are enabled iOS compatible images and videos will be saved in your device's gallery (Note: images and videos won't be saved in Offline)