Send message from Contacts -> Contact Info
Open chat room (If already exists) or create it with that contact.
The chat room will be presented and underneath we will change to the Chat tab. The back button item will indicate that you are going back to the “Chat”, not where you was when you tapped “Send Message” (Contact -> Contact Info).