Allow make an A/V call from Contacts app
MEGA will appear as a option on the Contacts app if the contact has an email added.
If the user is your contact in MEGA, you can make a call directly. To make it, a chat room is needed, so if the user is your contact in MEGA but you haven't a chat with him/her, the chat room will be created automatically before making the call.
If the user isn't your contact in MEGA, a view to invite him/her will appear, or a view to showing that "Invite Sent" in case is pending on the outgoing contact requests.

Refactor WarningTransferQuotaViewController to CustomModalAlertViewController to reuse this ViewController. Currently used for: Warning Transfer Quota, Invite Contact and Invite Sent