Refactored node details
This began as a bug task, but became a refactoring task because one bug was found after another, and also revealed several inconsistencies along the places this view is reused.
So this class has been simplified as much as possible, taking into account the complexity of it.
Some examples of the bugs encountered while working on this task are:

- Rows triggering wrong actions.
- Rows that should appear on some cases, were missing.
- Wrong actions show on different scenarios. This became worst when you entered on the details of folders inside incoming or outgoing shares.