Don't show the spinner in the chat list
Inside a chat room show "Connecting..." instead "No internet connection" if the connection state is Offline or InProgress

Previously the chat list was shown if all chat rooms wera connected. It may happen that the app can't connect to a one chat room and you weren't be able to see other chats that were connected.