Remove "generate_fa" app data
The SDK sets up the file attributes when a file without them is downloaded.
Previously, the code was added to avoid request the thumbnail and the preview. But, it is not happening currently, the app always requests the thumbnail to the servers and creates the preview from the original photo (the downloaded file), but the preview would be used only if the /tmp directory is deleted.