When receiving a message set it as seen only if the app is in the messages view controller.
Set last message as seen in messages view controller: in the methods viewWillAppear and app didBecomeActive.

If proceed, show the unread message label when opening the messages view controller from app in background and from other view controllers pushed/presented from messages view controller.

Tap on a notification: if the app is in the current messages view controller do nothing, if the app is in the group details of the current messages view controllers pop to the messages view controller, if the app is in other messages view controller close the chat room and pop to root view controller, then push to the messages view controller.

Make MEGAPhotoBrowserDelegate methods optionals