Transfer refactor
Due some concurrency issues discovered on the bugs 10774 and 10787, the main goal of this changes is to evolve the logic of the Transfer view to a dynamic one. So basically, in a very brief summary, while you are on the Transfers view (and don't tap on another segment) the transfers are updated with all the callbacks that drive them through the downloading or uploading process.
Several things have been simplified. One thing to highlight is that to ease the transition from a core data transfer to a MEGA one, now the appData of each transfer has the local identifier of the asset.