Share location - initial commit
Allow send a location (The user current location or a location selected in a map). Send the latitude, longitude and a base 64 image centered in a map 1000 x 1000 meters, the size of the image is 750x750 and the quality is 0.3.
Add compatibility with Geolocation messages.
Create the view and the model for Geolocation messages