Fix invalid binary in App Store Connect
Invalid Info.plist value - Your app can't claim an LSHandlerRank of "Owner" for any CFBundleDocumentTypes containing a public.data LSItemContentType. Please use "Alternate" as the LSHandlerRank instead.

LSHandlerRank: determines how Launch Services ranks this app among the apps that declare themselves editors or viewers of files of this type. The possible values are: Owner (this app is the primary creator of files of this type), Default (this app is an opener of files of this type; this value is also used if no rank is specified), Alternate (this app is a secondary viewer of files of this type), and None (this app is never selected to open files of this type, but it accepts drops of files of this type). Launch Services uses the value of LSHandlerRank to determine the app to use to open files of this type. The order of precedence is: Owner, Default, Alternate. This key is available in macOS 10.5 and later and iOS 3.0 and later.