Fix add/remove local/remote chat video delegate (11478)
The remote video delegates can't be removed correctly in the view will disappear because the session there is not valid. The same happens when adding them in the view will appear.
The session is not valid until the karere reports it with the delegates. So, it is not valid add / remove chat video delegate in the view will appear or view will disappear.
Call chat remove remote video delegate when the remote session is destroyed.
Call chat remove local video delegate when the user terminates the participation in the call.
Also, it is necessary remove the video delegates when navigating out ("minimize") the one on one call.