Avoid unnecessary remove remotes video delegates
didEndDisplayingCell only makes sense if have more than 6 peers in the call.
Remove all video listeners when hanging / hiding the call.
When change the call layout (and in the configure manual user on focus remove) the remote video if there is a peer selected.
In the remove all video listeners remove the focused remote if it is not hidden.
Improve the logs