When calling someone and enabling the speaker while it's ringing, when the call starts one has to re-enable the speaker button again to get the call on speaker
Use the mode AVAudioSessionModeVoiceChat for the AVAudioSession, this way the user can't mute the dialler tone.
We don't use AVAudioSessionModeVideoChat because in that mode is not possible change from Speaker to Receiver.

If there is a bluetooth device connected, when disabling the speaker the session route change with override reason is received. If there is not a bluetooth connected the notification with this change reason is not received.