Report incoming call or udpate a previous reported incoming call
Example: receive a call in a new chat (that you don't have in cache). The app reports the incoming call using the information received in the payload (chatid and callid), "connecting" is shown in the CallKit and later when the chat information is available (through onChatCallUpdate callback) update the information for the call previously reported.