Fix crash megachat::MegaChatCallHandler::onDestroy (top 1 in version 4.8.3 and 4.8.4)
The problem occurs when the app destroys MegaChatApi while a call is in progress. The call completion is asynchronous and ends up accessing the MegaChatCallHandler::onDestroy() implementation when the rtcModule has already been destroyed (accessed at removeCall()).
For now we solve it by avoiding that MegaChatApi deletion when the app is being terminated.