Fixes initialising the provider delegate
Need init the provider delegate after megaChatApi exists, if not the provider delegate is not added as a listener of megachatcalls. Example of error: someone calls you and hang the call, you still see the incoming call.
If logout release the provider delegate and the call manager
Init provider delegate when log-in