fix(core data): avoid leave messages in core data
If the application post a message and later the extensions doesn't receive the push, the messages posted
by the application will be persisted forever in the core data.
It doesn't make sense that the aplication post a notification and later the extension receives a push. if that
happens, probably we have a bug in the server logic for APNs (mutable-content).