15208: action over an incoming call that doesn't exist yet (#591)
* fix(CallKit): action over an incoming call that doesn't exist yet

Actions on a call, of which you have the chat id and call id but not the call itself (the MEGA call), should not fail.
Posible actions are: reject, answer and mute call.
Save the action and when you are connected to the chat and retrieve the call act accordingly

15208

* refactor(CallKit): private method and change log line

* refactor(calls): set up callId when the call is active

* fix(call): video show in remote but no in local

When responding to a remote call from which we do not have your call,
this call is reported as audio, so we should answer with audio only

The problem is:
1. Run the app with the delay to connect to chatd and leave the app in foreground
2. Make 1-to-1 video call from webclient
3. Answer the call on iOS
4. passes the delay and connects.
5. The webclient shows both videos but in iOS I only see the remote

* fix(callkit): report incoming call only if receive VoIP

Update information in onChatCallUpdate

Co-authored-by: Carlos Martin Acera <cma@mega.nz>