task: don't show incoming call notification when mic permission is "not authorised"
Before, the application showed an "Incoming call" notification when it received a VoIP
and the permission to access the microphone was "not authorised, instead reporting
incoming call to CallKit.
Now, with iOS 13 SDK, it is mandatory to report the incoming call. So, it doesn't make
sense to show the notification.