task: remove offline files alert when logout (11960)
* fix(myaccount): remove offline files alert when logout

Add hint label under logout button to indicate to the user the existence of transfers and/or offline files.
Removed the alert notifying the existence of files in offline after logout because the flow with PasswordReminder was not ideal.
Added new strings for hints and removed the one of the alert.
Removed the option of logout in [MEGALinkManager presentNode] if the node is not available for the session.

11960

* task: move check if exists offline files from helper to NSFileManager category

Remove unnecessary import

Co-authored-by: Javier Navarro <jn@mega.co.nz>