fix(NSE): messages with links show ((edited)) in the notification (15996)
When a push is received for a message of type contains meta rich preview,
retrieving the message sometimes indicates that it is of normal type and
sometimes of type contains meta rich preview. Sometimes the direct message
is reported (normal) and sometimes the editing of the message is reported
(normal -> rich preview). In any case, it should not show "Edited".